Initial Implementation
Currently involves a good amount of setup which I'd like to minimize
if possible. It does make it very configurable and a very simple
framework at this point which doesn't have to maintain much.

We provide a default `ObservableProperty` for very lightweight usage but would
recommend using this with an Rx implementation to get the full power of it.

Definitely needs docs an examples which I'd be happy to write more of
once we solidify the API a bit more and decide how much this lib needs
to do vs. using an Rx implementation.